Single sign-on and sign-off for the softwarefactory services
The services which rely on the LDAP authentication such as redmine
and gerrit are authenticated via a cookie which is created by the
authentication service when a user logs in to any of the service.

When a service is launched it will redirect the user to the central
authentication service which validates the user's credentials against
an LDAP server or through github servers via OAUTH and sets a cookie
that would be visible across all the services that operate on the
same 'top domain'

Similarly, when user logs out of any of the service, the request will
be redirected to the central server which signs out from all the services
through a series of redirection thereby unsetting all the cookies.

Right now, the central authentication service is designed to run inside
the managesf node but with its own hostname alias.

The functional tests are updated to absorb cookies based authentication.

Through this commit, there is no way we allow an anonymous user to access the
softwarefactory services.

Change-Id: Ia8d1f6cc1030f96beef568c1f96e5ed4a3e55635