Fix double top menu
This fixes the double top menu entry when using LDAP and Github.
Redirect URI are limited to Gerrit, Jenkins and Redmine, no matter what
URI is given as back argument.

A second fix is now included in the page itself; if the page itself is
nested it will remove escape the top iframe.

Closes-Bug: 14
Change-Id: I35135d84782acd8283a86d46d6d1424c87ec043b