Remove LDAP mocking in cauth and add local user authentication
This patches modifies the authentication and makes it possible to define
multiple local users independent from all other authentication sources.
Passwords are stored hashed and salted on the managesf node.

By default only the admin user is configured, and only if the top_domain equals
to tests.dom additional users are configured (for testing).

Change-Id: Ib7c313be0463c60940027ea67199ef3c395e4bd2